+++ src/accounts/internal/usecase/add-funds.go
@@ -2,15 +2,16 @@ package usecase
 
 import (
 	"marketplace/accounts/domain"
+
 	"github.com/go-pg/pg/v10"
 )
 
-type AddFundsCmd func (db *pg.DB, acc *domain.Account, fundsToAdd int64) (int64, error)
+type AddFundsCmd func (db *pg.DB, acc *domain.Account, fundsToAdd int64) (float64, error)
 
 func AddFunds() AddFundsCmd {
 
-	return func (db *pg.DB, acc *domain.Account, fundsToAdd int64 ) (int64, error) {
-		acc.Balance += fundsToAdd
+	return func (db *pg.DB, acc *domain.Account, fundsToAdd int64 ) (float64, error) {
+		acc.Balance += float64(fundsToAdd)
 
 		_, err := db.Model(acc).WherePK().Update()
